Added a MOTD server setting.
Message is shown to players upon logging in or registering. Nothing will show if the database setting is null, messages are also prefixed with "Message of the day" (added when the setting is loaded).

Message is stored as a global setting variable, will be expanded in the future. Database changes:
ALTER TABLE `settings` ADD `ServerMOTD` VARCHAR(64) NOT NULL AFTER `ModeName`;